Revert "Remove additional core data background context when downloading content of sections"
This reverts commit 8613d12e5e54de2d959218c9490178f29b26d3c4.

+++ iOS/Helpers/PersistenceManager/StreamPersistenceManager.swift
@@ -162,16 +162,22 @@ extension StreamPersistenceManager {
     func startDownloads(for section: CourseSection, options: [StreamPersistenceManager.Option] = []) -> Future<Void, XikoloError> {
         let promise = Promise<Void, XikoloError>()
 
-        self.persistentContainerQueue.addOperation {
-            let sectionDownloadFuture = section.items.compactMap { item in
-                return item.content as? Video
-            }.filter { video in
-                return self.downloadState(for: video) == .notDownloaded
-            }.map { video in
-                self.startDownload(for: video, options: options)
-            }.sequence().asVoid()
+        let sectionObjectID = section.objectID
 
-            promise.completeWith(sectionDownloadFuture)
+        self.persistentContainerQueue.addOperation {
+            let context = CoreDataHelper.persistentContainer.newBackgroundContext()
+            context.performAndWait {
+                guard let section: CourseSection = context.existingTypedObject(with: sectionObjectID) else { return }
+                let sectionDownloadFuture = section.items.compactMap { item in
+                    return item.content as? Video
+                }.filter { video in
+                    return self.downloadState(for: video) == .notDownloaded
+                }.map { video in
+                    self.startDownload(for: video, options: options)
+                }.sequence().asVoid()
+
+                promise.completeWith(sectionDownloadFuture)
+            }
         }
 
         return promise.future
@@ -181,14 +187,20 @@ extension StreamPersistenceManager {
     func deleteDownloads(for section: CourseSection) -> Future<Void, XikoloError> {
         let promise = Promise<Void, XikoloError>()
 
-        self.persistentContainerQueue.addOperation {
-            let sectionDeleteFuture = section.items.compactMap { item in
-                return item.content as? Video
-            }.map { video in
-                self.deleteDownload(for: video)
-            }.sequence().asVoid()
+        let sectionObjectID = section.objectID
 
-            return promise.completeWith(sectionDeleteFuture)
+        self.persistentContainerQueue.addOperation {
+            let context = CoreDataHelper.persistentContainer.newBackgroundContext()
+            context.performAndWait {
+                guard let section: CourseSection = context.existingTypedObject(with: sectionObjectID) else { return }
+                let sectionDeleteFuture = section.items.compactMap { item in
+                    return item.content as? Video
+                }.map { video in
+                    self.deleteDownload(for: video)
+                }.sequence().asVoid()
+
+                return promise.completeWith(sectionDeleteFuture)
+            }
         }
 
         return promise.future
